Move MdiContainerForm specific UIActions to new file.
